move navigation to library
